+++ cmd/presto/main.go
@@ -17,7 +17,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var version = "0.1.5"
+var version = "0.1.6"
 var verbose bool
 
 func main() {